open from clipboard
