+++ src/Captura/Captura.csproj
@@ -76,9 +76,17 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="ImageEditorWindow.xaml.cs">
+    <Compile Include="ImageEditor\HistoryItem.cs" />
+    <Compile Include="ImageEditor\HistoryState.cs" />
+    <Compile Include="ImageEditor\ImageEditorViewModel.cs" />
+    <Compile Include="ImageEditor\ImageEditorWindow.xaml.cs">
       <DependentUpon>ImageEditorWindow.xaml</DependentUpon>
     </Compile>
+    <Compile Include="ImageEditor\ImageEffect.cs" />
+    <Compile Include="ImageEditor\ImageFromClipboard.cs" />
+    <Compile Include="ImageEditor\ModifyPixel.cs" />
+    <Compile Include="ImageEditor\PixelFunctionFactory.cs" />
+    <Compile Include="ImageEditor\StrokeHistory.cs" />
     <Compile Include="Models\HitType.cs" />
     <Compile Include="Presentation\Controls\LayerFrame.xaml.cs">
       <DependentUpon>LayerFrame.xaml</DependentUpon>
@@ -185,7 +193,7 @@
       <DependentUpon>AboutView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Presentation\WpfExtensions.cs" />
-    <Page Include="ImageEditorWindow.xaml">
+    <Page Include="ImageEditor\ImageEditorWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
