+++ src/Captura/ImageEditor/HistoryItem.cs
@@ -1,0 +1,4 @@
+﻿namespace Captura
+{
+    public abstract class HistoryItem { }
+}
