+++ src/Captura/ImageEditor/ImageEditorViewModel.cs
@@ -3,132 +3,14 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Ink;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
+using Captura.Models;
 using Microsoft.Win32;
 
 namespace Captura
 {
-    public enum ImageEffect
-    {
-        None,
-        Negative,
-        Green,
-        Red,
-        Blue,
-        Grayscale,
-        Sepia
-    }
-
-    public delegate void ModifyPixel(ref byte Red, ref byte Green, ref byte Blue);
-
-    public abstract class HistoryItem { }
-
-    public class HistoryState : HistoryItem
-    {
-        public ImageEffect Effect { get; set; }
-
-        public int Brightness { get; set; }
-
-        public int Contrast { get; set; }
-
-        public int Rotation { get; set; }
-
-        public bool FlipX { get; set; }
-
-        public bool FlipY { get; set; }
-    }
-
-    public class StrokeHistory : HistoryItem
-    {
-        public int EditingOperationCount { get; set; }
-
-        public InkCanvasEditingMode EditingMode { get; set; }
-
-        public List<Stroke> Added { get; } = new List<Stroke>();
-        public List<Stroke> Removed { get; } = new List<Stroke>();
-    }
-
-    public static class PixelFunctionFactory
-    {
-        static void Negative(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            Red = (byte)(255 - Red);
-            Green = (byte)(255 - Green);
-            Blue = (byte)(255 - Blue);
-        }
-
-        static void Green(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            Red = Blue = 0;
-        }
-
-        static void Red(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            Green = Blue = 0;
-        }
-
-        static void Blue(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            Red = Green = 0;
-        }
-
-        static void Grayscale(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            var pixel = 0.299 * Red + 0.587 * Green + 0.114 * Blue;
-
-            if (pixel > 255)
-                pixel = 255;
-
-            Red = Green = Blue = (byte)pixel;
-        }
-
-        static void Sepia(ref byte Red, ref byte Green, ref byte Blue)
-        {
-            var newRed = 0.393 * Red + 0.769 * Green + 0.189 * Blue;
-            var newGreen = 0.349 * Red + 0.686 * Green + 0.168 * Blue;
-            var newBlue = 0.272 * Red + 0.534 * Green + 0.131 * Blue;
-
-            // Red
-            Red = (byte)(newRed > 255 ? 255 : newRed);
-
-            // Green
-            Green = (byte)(newGreen > 255 ? 255 : newGreen);
-
-            // Blue
-            Blue = (byte)(newBlue > 255 ? 255 : newBlue);
-        }
-
-        public static ModifyPixel GetEffectFunction(ImageEffect Effect)
-        {
-            switch (Effect)
-            {
-                case ImageEffect.Negative:
-                    return Negative;
-
-                case ImageEffect.Blue:
-                    return Blue;
-
-                case ImageEffect.Green:
-                    return Green;
-
-                case ImageEffect.Red:
-                    return Red;
-
-                case ImageEffect.Sepia:
-                    return Sepia;
-
-                case ImageEffect.Grayscale:
-                    return Grayscale;
-
-                default:
-                    return null;
-            }
-        }
-    }
-
     public class ImageEditorViewModel : NotifyPropertyChanged
     {
         int _stride;
@@ -143,6 +25,7 @@ namespace Captura
         const int ContrastStep = 10;
 
         public ICommand OpenCommand { get; }
+        public ICommand OpenFromClipboardCommand { get; }
         public DelegateCommand UndoCommand { get; }
         public DelegateCommand RedoCommand { get; }
 
@@ -158,6 +41,7 @@ namespace Captura
         public ImageEditorViewModel()
         {
             OpenCommand = new DelegateCommand(Open);
+            OpenFromClipboardCommand = new DelegateCommand(OpenFromClipboard);
             UndoCommand = new DelegateCommand(Undo, false);
             RedoCommand = new DelegateCommand(Redo, false);
 
@@ -253,9 +137,9 @@ namespace Captura
             }, false);
         }
 
-        BitmapFrame _originalBmp;
+        BitmapSource _originalBmp;
 
-        public BitmapFrame OriginalBitmap
+        public BitmapSource OriginalBitmap
         {
             get => _originalBmp;
             private set
@@ -425,8 +309,12 @@ namespace Captura
             Rotation = 0;
             FlipX = FlipY = false;
 
+            InkCanvas.Strokes.Clear();
+
             _undoStack.Clear();
+            _redoStack.Clear();
             UndoCommand.RaiseCanExecuteChanged(false);
+            RedoCommand.RaiseCanExecuteChanged(false);
         }
 
         void Open()
@@ -440,30 +328,49 @@ namespace Captura
 
             if (ofd.ShowDialog().GetValueOrDefault())
             {
-                Reset();
-
                 var decoder = BitmapDecoder.Create(new Uri(ofd.FileName),
                     BitmapCreateOptions.None, BitmapCacheOption.None);
 
-                OriginalBitmap = decoder.Frames[0];
-
-                _stride = OriginalBitmap.PixelWidth * (OriginalBitmap.Format.BitsPerPixel / 8);
-
-                _data = new byte[_stride * OriginalBitmap.PixelHeight];
-
-                EditedBitmap = new WriteableBitmap(OriginalBitmap);
-                
-                UpdateTransformBitmap();
-
-                SetEffectCommand.RaiseCanExecuteChanged(true);
-                SetBrightnessCommand.RaiseCanExecuteChanged(true);
-                SetContrastCommand.RaiseCanExecuteChanged(true);
-
-                RotateRightCommand.RaiseCanExecuteChanged(true);
-                RotateLeftCommand.RaiseCanExecuteChanged(true);
-                FlipXCommand.RaiseCanExecuteChanged(true);
-                FlipYCommand.RaiseCanExecuteChanged(true);
+                OnOpen(decoder.Frames[0]);
             }
+        }
+
+        void OpenFromClipboard()
+        {
+            var img = ImageFromClipboard.Get();
+
+            if (img == null)
+            {
+                ServiceProvider.Get<MessageProvider>().ShowError("No Image on Clipboard");
+
+                return;
+            }
+
+            OnOpen(img);
+        }
+
+        void OnOpen(BitmapSource Frame)
+        {
+            Reset();
+
+            OriginalBitmap = Frame;
+
+            _stride = OriginalBitmap.PixelWidth * (OriginalBitmap.Format.BitsPerPixel / 8);
+
+            _data = new byte[_stride * OriginalBitmap.PixelHeight];
+
+            EditedBitmap = new WriteableBitmap(OriginalBitmap);
+
+            UpdateTransformBitmap();
+
+            SetEffectCommand.RaiseCanExecuteChanged(true);
+            SetBrightnessCommand.RaiseCanExecuteChanged(true);
+            SetContrastCommand.RaiseCanExecuteChanged(true);
+
+            RotateRightCommand.RaiseCanExecuteChanged(true);
+            RotateLeftCommand.RaiseCanExecuteChanged(true);
+            FlipXCommand.RaiseCanExecuteChanged(true);
+            FlipYCommand.RaiseCanExecuteChanged(true);
         }
 
         public void Save(BitmapSource Bmp)
@@ -644,144 +551,6 @@ namespace Captura
         public void IncrementEditingOperationCount()
         {
             ++_editingOperationCount;
-        }
-    }
-
-    public partial class ImageEditorWindow
-    {
-        public ImageEditorWindow()
-        {
-            InitializeComponent();
-
-            if (DataContext is ImageEditorViewModel vm)
-            {
-                vm.PropertyChanged += (S, E) =>
-                {
-                    if (E.PropertyName == nameof(vm.TransformedBitmap))
-                        UpdateInkCanvas();
-                };
-
-                vm.InkCanvas = InkCanvas;
-
-                InkCanvas.Strokes.StrokesChanged += (S, E) =>
-                {
-                    var item = new StrokeHistory();
-                    
-                    item.Added.AddRange(E.Added);
-                    item.Removed.AddRange(E.Removed);
-
-                    vm.AddInkHistory(item);
-                };
-            }
-
-            Image.SizeChanged += (S, E) => UpdateInkCanvas();
-
-            ColorPicker.SelectedColor = Color.FromRgb(27, 27, 27);
-            ModesBox.SelectedIndex = 0;
-            SizeBox.Value = 10;
-
-            InkCanvas.DefaultDrawingAttributes.FitToCurve = true;
-        }
-
-        void Exit(object Sender, RoutedEventArgs E)
-        {
-            Close();
-        }
-
-        void SizeBox_OnValueChanged(object Sender, RoutedPropertyChangedEventArgs<object> E)
-        {
-            if (InkCanvas != null && E.NewValue is int i)
-                InkCanvas.DefaultDrawingAttributes.Height = InkCanvas.DefaultDrawingAttributes.Width = i;
-        }
-
-        void ModesBox_OnSelectionChanged(object Sender, SelectionChangedEventArgs E)
-        {
-            if (ModesBox.SelectedValue is InkCanvasEditingMode mode)
-            {
-                InkCanvas.EditingMode = mode;
-
-                if (mode == InkCanvasEditingMode.Ink)
-                {
-                    InkCanvas.UseCustomCursor = true;
-                    InkCanvas.Cursor = Cursors.Pen;
-                }
-                else InkCanvas.UseCustomCursor = false;
-            }
-        }
-
-        void ColorPicker_OnSelectedColorChanged(object Sender, RoutedPropertyChangedEventArgs<Color?> E)
-        {
-            if (E.NewValue != null && InkCanvas != null)
-                InkCanvas.DefaultDrawingAttributes.Color = E.NewValue.Value;
-        }
-
-        void UpdateInkCanvas()
-        {
-            if (DataContext is ImageEditorViewModel vm && vm.TransformedBitmap != null && Image.ActualWidth > 0)
-            {
-                InkCanvas.IsEnabled = true;
-
-                InkCanvas.Width = vm.OriginalBitmap.PixelWidth;
-                InkCanvas.Height = vm.OriginalBitmap.PixelHeight;
-
-                var rotate = new RotateTransform(vm.Rotation, vm.OriginalBitmap.PixelWidth / 2.0, vm.OriginalBitmap.PixelHeight / 2.0);
-
-                var tilted = Math.Abs(vm.Rotation / 90) % 2 == 1;
-                
-                var scale = new ScaleTransform(
-                    ((tilted ? Image.ActualHeight : Image.ActualWidth) / InkCanvas.Width) * (vm.FlipX ? -1 : 1),
-                    ((tilted ? Image.ActualWidth : Image.ActualHeight) / InkCanvas.Height) * (vm.FlipY ? -1 : 1)
-                );
-
-                InkCanvas.LayoutTransform = new TransformGroup
-                {
-                    Children =
-                    {
-                        rotate,
-                        scale
-                    }
-                };
-            }
-        }
-
-        void OnSave(object Sender, ExecutedRoutedEventArgs E)
-        {
-            if (DataContext is ImageEditorViewModel vm && vm.TransformedBitmap != null)
-            {
-                var drawingVisual = new DrawingVisual();
-
-                var copy = vm.EditedBitmap;
-                var transform = vm.TransformedBitmap.Transform;
-
-                using (var drawingContext = drawingVisual.RenderOpen())
-                {
-                    drawingContext.DrawImage(copy, new Rect(0, 0, copy.Width, copy.Height));
-
-                    InkCanvas.Strokes.Draw(drawingContext);
-
-                    drawingContext.Close();
-
-                    var bitmap = new RenderTargetBitmap((int)copy.Width,
-                        (int)copy.Height,
-                        copy.DpiX,
-                        copy.DpiY,
-                        PixelFormats.Pbgra32);
-
-                    bitmap.Render(drawingVisual);
-
-                    var transformedRendered = new TransformedBitmap(bitmap, transform);
-
-                    vm.Save(transformedRendered);
-                }
-            }
-        }
-
-        void InkCanvas_OnMouseUp(object Sender, MouseButtonEventArgs E)
-        {
-            if (DataContext is ImageEditorViewModel vm)
-            {
-                vm.IncrementEditingOperationCount();
-            }
         }
     }
 }
