+++ src/Captura/ImageEditor/ImageEffect.cs
@@ -1,0 +1,13 @@
+﻿namespace Captura
+{
+    public enum ImageEffect
+    {
+        None,
+        Negative,
+        Green,
+        Red,
+        Blue,
+        Grayscale,
+        Sepia
+    }
+}
