+++ src/Captura/ImageEditor/ModifyPixel.cs
@@ -1,0 +1,4 @@
+﻿namespace Captura
+{
+    public delegate void ModifyPixel(ref byte Red, ref byte Green, ref byte Blue);
+}
